requirements.txt
requests
python-dotenv
text-unidecode
tqdm
pandas
